Tweeking margin before tag category change
I realized that the lack of padding before the first tag of each category was not universal, but instead something I was only observing in Safari mobile. Turns out it was caused by the fact that browser doesn't allow font-size to be so small it makes text invisible, so I changed my strategy and added a whitespace before the category change which I've resized so it should only have an effect when needed. Also cleaned up the selectors for this rule.