Fixing IE bug
After the font size was shrunk to 0.1%, Internet Explorer set it to 0pt and did not allow it to be increased in the child elements. By changing the values to 5% and 2000%, the same vanishing effect should be preserved, but IE does allow the child elements' contents to be resized back to normal.